test book 3/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Pillow==8.3.1
email-validator==1.1.3
python-dotenv==0.19.0
Werkzeug==2.0.1